Podcast builder (#1)
I want to have a private* podcast for the reading plan.

This PR introduces some new tooling that creates an audio file for every weekday, as well as a script that generates the podcast feed XML file.

I've used the audio file generation tooling to generate all the audio files for the year, and I've stored them in a GCS bucket.

I've drafted a GitHub Action script that should be able to run the (much faster and less resource-intensive) script for generating the podcast feed, which I want to do automatically every weekday, and then uploading it to the GCS bucket.

*I will have to review the ESV API Terms carefully - would be great to make this public! But for now, the source code is at least public.